Change to initial fetch by server components
