+++ src/app/ListView.tsx
@@ -1,7 +1,6 @@
 "use client"
 import Image from "next/image"
 import Link from "next/link"
-import React from "react"
 import { NFTCollectionType } from "src/shared/NFTCollection.type"
 
 type ListViewProps = {
