+++ src/app/getActionApi.ts
@@ -1,7 +1,9 @@
 import axios, { AxiosRequestConfig, AxiosResponse, AxiosError } from 'axios';
 import { ApiPath, ApiReqBodies, ApiResBodies } from 'src/shared/apiTypes';
 
-const API_URL = process.env.NEXT_PUBLIC_API_URL;
+const isServer = typeof window === 'undefined';
+let API_URL: string | undefined;
+API_URL = isServer ? process.env.SERVER_SIDE_API_URL : process.env.NEXT_PUBLIC_CLIENT_SIDE_API_URL;
 
 function createAxiosInstance() {
   const config: AxiosRequestConfig = { baseURL: API_URL }
