+++ src/app/layout.tsx
@@ -1,10 +1,11 @@
 import Link from 'next/link'
 import './globals.css'
 import { Inter } from 'next/font/google'
+import { Metadata } from 'next';
 
 const inter = Inter({ subsets: ['latin'] })
 
-export const metadata = {
+export const metadata: Metadata = {
   title: 'NFT Analytics Site',
   description: 'Generated by konpay.eth',
 }
